+++ frontend/src/components/Carousel/components/CarouselItem/index.scss
