+++ frontend/src/components/Carousel/components/CarouselItem/type.ts
@@ -1,0 +1,11 @@
+import { CSSProperties } from "react";
+
+export interface ICarouselItem {
+  titleList?: string[];
+  descList?: string[];
+  style?: CSSProperties;
+}
+
+export interface ICarouselItemProps {
+  data?: ICarouselItem;
+}
