+++ frontend/src/components/Carousel/components/TimeBarItem/index.scss
