+++ frontend/src/components/Carousel/components/TimeBarItem/type.ts
@@ -1,0 +1,4 @@
+export interface ITimeBarItemProps {
+  active?: boolean;
+  stayTime?: number;
+}
