+++ frontend/src/components/Carousel/type.ts
@@ -1,0 +1,7 @@
+import { ICarouselItem } from "./components/CarouselItem/type";
+
+export interface ICarouselProps {
+  list?: ICarouselItem[];
+  stayTime?: number;
+  changeTime?: number;
+}
